refactor: refactor create generic
