+++ internal/cli/cluster/update.go
@@ -24,6 +24,7 @@ import (
 	apiconst "github.com/oceanbase/ob-operator/api/constants"
 	"github.com/oceanbase/ob-operator/api/types"
 	"github.com/oceanbase/ob-operator/api/v1alpha1"
+	"github.com/oceanbase/ob-operator/internal/cli/generic"
 	oceanbaseconst "github.com/oceanbase/ob-operator/internal/const/oceanbase"
 	"github.com/oceanbase/ob-operator/internal/dashboard/business/constant"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/common"
@@ -31,7 +32,7 @@ import (
 )
 
 type UpdateOptions struct {
-	ResourceOptions
+	generic.ResourceOptions
 	Resource     common.ResourceSpec             `json:"resource"`
 	Storage      *param.OBServerStorageSpec      `json:"storage"`
 	UpdateType   string                          `json:"updateType"`
